Update logic of existing input number filter classes * Add a class CustomNumFilter that ensures only number is present on the field the filter is applied to * Use the custom CustomNumFilter for NumInputFilter (old) or now NumLimFilter (new) class * Shift the two input filter classes to a separate package (app.grapheneos.camera.inputfilter)
